feat: move tests setup file to src/tests
