+++ packages/web/components/navbar-osmosis-update/__tests__/navbar-osmosis-update.spec.tsx
@@ -6,7 +6,7 @@ import { resetLDMocks } from "jest-launchdarkly-mock";
 import { rest } from "msw";
 
 import NavbarOsmosisUpdate from "~/components/navbar-osmosis-update";
-import { server } from "~/utils/msw";
+import { server } from "~/tests/msw";
 import { mockFeatureFlags, renderWithProviders } from "~/utils/test-utils";
 
 beforeEach(() => {
